Adding ESM webworker support
Classic webworkers support importScripts() however ESM ones do not

Have opted to use dynamic imports (https://web.dev/articles/module-workers). This might not always behave as expected if relying on global variables to be populated by the scripts being imported. In that case; we would need to look at the result of the promise and use those variables returned.